take confirmation before delete

+++ admin/src/pages/Complaint/Complaint.jsx
@@ -65,6 +65,8 @@ const Complaint = () => {
   };
 
   const handleDelete = async (id) => {
+    const confirmed = window.confirm("Are you sure you want to permanantly delete this complaint?");
+    if (!confirmed) return;
     await deleteData(`${singleDeleteUrl}/${id}`, validToken);
   };
 
@@ -131,7 +133,7 @@ const Complaint = () => {
             ) : (
               <tr>
                 <td colSpan="7" className="text-center">
-                  No Data.
+                  No Data
                 </td>
               </tr>
             )
