+++ admin/src/pages/Flat/Flat.jsx
@@ -56,6 +56,8 @@ const Flat = () => {
   };
 
   const handleDelete = async (id) => {
+    const confirmed = window.confirm("Are you sure you want to permanantly delete this flat?");
+    if (!confirmed) return;
     await deleteData(`${singleDeleteUrl}/${id}`, validToken);
   };
 
@@ -109,7 +111,7 @@ const Flat = () => {
             ) : (
               <tr>
                 <td colSpan="7" className="text-center">
-                  No Data.
+                  No Data
                 </td>
               </tr>
             )
